requirements.txt
websockets>=12.0
httpx>=0.27
python-dotenv>=1.0
orjson>=3.9.0
prometheus-client>=0.19.0
